backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
pandas>=2.0.0
beancount>=2.3.0
pyyaml>=6.0
